information_portal_backend/urls.py
"""information_portal_backend URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/3.1/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path
from django.conf.urls import url
from django.conf.urls import include
from django.conf.urls.static import static

from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi

from information_portal_backend.settings import MEDIA_ROOT, MEDIA_URL 

schema_view = get_schema_view(
   openapi.Info(
      title="Information Portal API",
      default_version='v1',
      description="Test description",
      contact=openapi.Contact(email="contact@snippets.local"),
      license=openapi.License(name="BSD License"),
   ),
   public=False,
   permission_classes=(permissions.IsAuthenticated,),
)

urlpatterns = [
    path('admin/', admin.site.urls),

    # APIs
    url(r'^api/', include( 'category.urls' )),
    url(r'^api/', include( 'content.urls' )),
    url(r'^api/', include( 'homepage.urls' )),

    # docs
    url(r'^docs/$', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),

] + static(MEDIA_URL, document_root=MEDIA_ROOT)

admin.site.site_header = 'Information Portal Admin Panel'
admin.site.site_title = 'Information Portal'
